LPS-105599 Move fragment configuration field components to floating-toolbar directory
