+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/floating-toolbar/ColorPaletteField.js
@@ -15,7 +15,7 @@
 import ClayForm from '@clayui/form';
 import React from 'react';
 
-import ColorPalette from '../../common/components/ColorPalette';
+import ColorPalette from '../../../common/components/ColorPalette';
 
 export const ColorPaletteField = ({field, onValueSelect, value}) => (
 	<ClayForm.Group small>
